+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Azure Credentials Plugin Changelog
 
+## Version 3.0.0, 2020-03-30
+* Upgrade Azure-SDK dependency to Azure-security-keyvault-secrets
+* Fix compatibility issue with azure-keyvault
+* Remove UsernamePassword support in Azure Key Vault
+* Remove UsernamePassword support in ```AzureClient``` creation
+
+## Version 2.0.2, 2020-03-24
+* Upgrade Azure SDK dependency to version 1.31.0
+
 ## Version 2.0.1, 2020-01-25
 * Fix service principal secret for key vault credential